requirements.txt
### Use this .txt to version the required dependencies to run the project.
### Note that every line that begins with a # is ignored by the file (commented row).
### Please do not change the dependencies that are already set here.

# Testing
tox==3.24.3

# AWS
boto3==1.18.36

# CLI
fire==0.4.0

# API
flask==2.0.2
gunicorn==20.1.0

# Jupyter Notebook Kernel
jupyter==1.0.0

# Data Analysis
pandas==1.3.4

# Environment Variables
python-dotenv==0.19.2

# YAML Converter
strictyaml==1.4.4
ruamel.yaml==0.16.12

# Typing
pydantic==1.8.2

# Scikit-Learn
scikit-learn==1.0.1
